% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arReceivable.R
\name{arReceivable_delete}
\alias{arReceivable_delete}
\title{清空临时表}
\usage{
arReceivable_delete(dms_token)
}
\arguments{
\item{dms_token}{第二个参数}
}
\value{
两个数的和
}
\description{
清空临时表
}
\examples{
arReceivable_delete
}
