% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arReceivable.R
\name{arReceivable_insert}
\alias{arReceivable_insert}
\title{更新list表和表头表体数据}
\usage{
arReceivable_insert(dms_token)
}
\arguments{
\item{dms_token}{第二个参数}
}
\value{
两个数的和
}
\description{
更新list表和表头表体数据
}
\examples{
arReceivable_insert
}
