% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arReceivable.R
\name{arReceivable_select}
\alias{arReceivable_select}
\title{查询list表}
\usage{
arReceivable_select(dms_token, FStartDate, FEndDate)
}
\arguments{
\item{dms_token}{第二个参数}

\item{FEndDate}{}
}
\value{
两个数的和
}
\description{
查询list表
}
\examples{
arReceivable_select
}
